chore(server): core IoC 1 - graph/directives
